--- components/mobile-title.tsx
@@ -1,11 +1,0 @@
-import TitleText from "./title-text";
-
-export default function MobileTitleComponent() {
-    return (
-        <div className="w-full h-fit flex items-center justify-center">
-            <div className="w-fit mt-16">
-                <TitleText />
-            </div>
-        </div>
-    );
-}
