+++ components/nav-bar/nav-bar.tsx
@@ -6,7 +6,7 @@ export default function NavBar() {
 		<div className="bg-hackrpi-secondary-yellow w-full h-12">
 			<div className="flex justify-start items-center h-full">
 				<NavBarLink href="/">HackRPI logo here</NavBarLink>
-                <NavBarLink href="/sponsor-us">Sponsor Us</NavBarLink>
+				<NavBarLink href="/sponsor-us">Sponsor Us</NavBarLink>
 			</div>
 		</div>
 	);
